添加蓝牙动态添加点数据，以及蓝牙通讯问题

+++ app/src/main/java/com/example/customview/mp_chart/DynamicLineChartManager.java
@@ -146,7 +146,7 @@ public class DynamicLineChartManager {
      *
      * @param number
      */
-    public void addEntry(int number) {
+    public void addEntry(float number) {
 
         //最开始的时候才添加 lineDataSet（一个lineDataSet 代表一条线）
         if (lineDataSet.getEntryCount() == 0) {
